Move GlyphName to names.rs
